Extract HTML leaderboard generation to dedicated module
